merge: 合并代码，解决代码冲突

+++ src/main/java/io/github/talelin/latticy/service/impl/AdminServiceImpl.java
@@ -162,6 +162,10 @@ public class AdminServiceImpl implements AdminService {
             throw new ForbiddenException(10075);
         }
         throwGroupNotExistById(id);
+        List<Integer> groupUserIds = groupService.getGroupUserIds(id);
+        if(groupUserIds.size() > 0) {
+            throw new ForbiddenException(10027);
+        }
         return groupService.removeById(id);
     }
